+++ src/commands/mod.rs
@@ -1,8 +1,8 @@
 //! This module implements the command-line API.
 
-use crate::lib::{AnyhowResult, AuthInfo};
+use crate::{get_auth, lib::AnyhowResult, BaseOpts};
 use anyhow::Context;
-use clap::Parser;
+use clap::{Args, Parser};
 use std::io::{self, Write};
 use tokio::runtime::Runtime;
 
@@ -28,66 +28,88 @@ pub use public::get_ids;
 #[derive(Parser)]
 pub enum Command {
     /// Prints the principal id and the account id.
-    PublicIds(public::PublicOpts),
-    Send(send::SendOpts),
-    Transfer(transfer::TransferOpts),
+    PublicIds(BaseOpts<public::PublicOpts>),
+    Send(BaseOpts<send::SendOpts>),
+    Transfer(BaseOpts<transfer::TransferOpts>),
     /// Claim seed neurons from the Genesis Token Canister.
-    ClaimNeurons,
-    NeuronStake(neuron_stake::StakeOpts),
-    NeuronManage(neuron_manage::ManageOpts),
+    ClaimNeurons(BaseOpts<Empty>),
+    NeuronStake(BaseOpts<neuron_stake::StakeOpts>),
+    NeuronManage(BaseOpts<neuron_manage::ManageOpts>),
     /// Signs the query for all neurons belonging to the signing principal.
-    ListNeurons(list_neurons::ListNeuronsOpts),
-    ListProposals(list_proposals::ListProposalsOpts),
-    GetProposalInfo(get_proposal_info::GetProposalInfoOpts),
-    GetNeuronInfo(get_neuron_info::GetNeuronInfoOpts),
+    ListNeurons(BaseOpts<list_neurons::ListNeuronsOpts>),
+    ListProposals(BaseOpts<list_proposals::ListProposalsOpts>),
+    GetProposalInfo(BaseOpts<get_proposal_info::GetProposalInfoOpts>),
+    GetNeuronInfo(BaseOpts<get_neuron_info::GetNeuronInfoOpts>),
     /// Queries a ledger account balance.
-    AccountBalance(account_balance::AccountBalanceOpts),
+    AccountBalance(BaseOpts<account_balance::AccountBalanceOpts>),
     /// Update node provider details
-    UpdateNodeProvider(update_node_provider::UpdateNodeProviderOpts),
-    ReplaceNodeProviderId(replace_node_provide_id::ReplaceNodeProviderIdOpts),
+    UpdateNodeProvider(BaseOpts<update_node_provider::UpdateNodeProviderOpts>),
+    ReplaceNodeProviderId(BaseOpts<replace_node_provide_id::ReplaceNodeProviderIdOpts>),
     /// Generate a mnemonic seed phrase and generate or recover PEM.
-    Generate(generate::GenerateOpts),
+    Generate(BaseOpts<generate::GenerateOpts>),
     /// Print QR Scanner dapp QR code: scan to start dapp to submit QR results.
     ScannerQRCode,
     /// Print QR code for data e.g. principal id.
-    QRCode(qrcode::QRCodeOpts),
+    QRCode(BaseOpts<qrcode::QRCodeOpts>),
 }
 
-pub fn exec(auth: &AuthInfo, qr: bool, fetch_root_key: bool, cmd: Command) -> AnyhowResult {
+#[derive(Args)]
+pub struct Empty;
+
+pub fn dispatch(cmd: Command) -> AnyhowResult {
     let runtime = Runtime::new().expect("Unable to create a runtime");
     match cmd {
-        Command::PublicIds(opts) => public::exec(auth, opts),
-        Command::Transfer(opts) => transfer::exec(auth, opts).and_then(|out| print_vec(qr, &out)),
+        Command::PublicIds(opts) => public::exec(&get_auth(opts.global_opts)?, opts.command_opts)?,
+        Command::Transfer(opts) => {
+            let qr = opts.global_opts.qr;
+            let out = transfer::exec(&get_auth(opts.global_opts)?, opts.command_opts)?;
+            print_vec(qr, &out)?;
+        }
         Command::NeuronStake(opts) => {
-            neuron_stake::exec(auth, opts).and_then(|out| print_vec(qr, &out))
+            let qr = opts.global_opts.qr;
+            let out = neuron_stake::exec(&get_auth(opts.global_opts)?, opts.command_opts)?;
+            print_vec(qr, &out)?;
         }
         Command::NeuronManage(opts) => {
-            neuron_manage::exec(auth, opts).and_then(|out| print_vec(qr, &out))
+            let qr = opts.global_opts.qr;
+            let out = neuron_manage::exec(&get_auth(opts.global_opts)?, opts.command_opts)?;
+            print_vec(qr, &out)?;
         }
         Command::ListNeurons(opts) => {
-            list_neurons::exec(auth, opts).and_then(|out| print_vec(qr, &out))
+            let qr = opts.global_opts.qr;
+            let out = list_neurons::exec(&get_auth(opts.global_opts)?, opts.command_opts)?;
+            print_vec(qr, &out)?;
         }
-        Command::ClaimNeurons => claim_neurons::exec(auth).and_then(|out| print_vec(qr, &out)),
-        Command::ListProposals(opts) => {
-            runtime.block_on(async { list_proposals::exec(opts, fetch_root_key).await })
+        Command::ClaimNeurons(opts) => {
+            let qr = opts.global_opts.qr;
+            claim_neurons::exec(&get_auth(opts.global_opts)?)
+                .and_then(|out| print_vec(qr, &out))?;
         }
-        Command::GetProposalInfo(opts) => {
-            runtime.block_on(async { get_proposal_info::exec(opts, fetch_root_key).await })
-        }
-        Command::GetNeuronInfo(opts) => {
-            runtime.block_on(async { get_neuron_info::exec(opts, fetch_root_key).await })
-        }
-        Command::AccountBalance(opts) => {
-            runtime.block_on(async { account_balance::exec(opts, fetch_root_key).await })
-        }
+        Command::ListProposals(opts) => runtime.block_on(async {
+            list_proposals::exec(opts.command_opts, opts.global_opts.fetch_root_key).await
+        })?,
+        Command::GetProposalInfo(opts) => runtime.block_on(async {
+            get_proposal_info::exec(opts.command_opts, opts.global_opts.fetch_root_key).await
+        })?,
+        Command::GetNeuronInfo(opts) => runtime.block_on(async {
+            get_neuron_info::exec(opts.command_opts, opts.global_opts.fetch_root_key).await
+        })?,
+        Command::AccountBalance(opts) => runtime.block_on(async {
+            account_balance::exec(opts.command_opts, opts.global_opts.fetch_root_key).await
+        })?,
         Command::UpdateNodeProvider(opts) => {
-            update_node_provider::exec(auth, opts).and_then(|out| print(&out))
+            let out = update_node_provider::exec(&get_auth(opts.global_opts)?, opts.command_opts)?;
+            print(&out)?;
         }
         Command::ReplaceNodeProviderId(opts) => {
-            replace_node_provide_id::exec(auth, opts).and_then(|out| print(&out))
+            let out =
+                replace_node_provide_id::exec(&get_auth(opts.global_opts)?, opts.command_opts)?;
+            print(&out)?;
         }
-        Command::Send(opts) => runtime.block_on(async { send::exec(opts, fetch_root_key).await }),
-        Command::Generate(opts) => generate::exec(opts),
+        Command::Send(opts) => runtime.block_on(async {
+            send::exec(opts.command_opts, opts.global_opts.fetch_root_key).await
+        })?,
+        Command::Generate(opts) => generate::exec(opts.command_opts)?,
         // QR code for URL: https://p5deo-6aaaa-aaaab-aaaxq-cai.raw.ic0.app/
         // Source code: https://github.com/ninegua/ic-qr-scanner
         Command::ScannerQRCode => {
@@ -113,10 +135,10 @@ pub fn exec(auth: &AuthInfo, qr: bool, fetch_root_key: bool, cmd: Command) -> An
 █████████████████████████████████████
 █████████████████████████████████████"
             );
-            Ok(())
         }
-        Command::QRCode(opts) => qrcode::exec(opts),
+        Command::QRCode(opts) => qrcode::exec(opts.command_opts)?,
     }
+    Ok(())
 }
 
 // Using println! for printing to STDOUT and piping it to other tools leads to