+++ tests/commands/list-neurons-many.sh
@@ -1,1 +1,1 @@
-${CARGO_TARGET_DIR:-../target}/debug/quill --pem-file - list-neurons 123 456 789 | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -
+${CARGO_TARGET_DIR:-../target}/debug/quill list-neurons 123 456 789 --pem-file - | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -