+++ tests/commands/neuron-manage-add-hot-key.sh
@@ -1,1 +1,1 @@
-${CARGO_TARGET_DIR:-../target}/debug/quill --pem-file - neuron-manage 2313380519530470538 --add-hot-key fdsgv-62ihb-nbiqv-xgic5-iefsv-3cscz-tmbzv-63qd5-vh43v-dqfrt-pae | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -
+${CARGO_TARGET_DIR:-../target}/debug/quill neuron-manage 2313380519530470538 --pem-file - --add-hot-key fdsgv-62ihb-nbiqv-xgic5-iefsv-3cscz-tmbzv-63qd5-vh43v-dqfrt-pae | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -