+++ tests/commands/neuron-manage-disburse-stop-dissolving.sh
@@ -1,1 +1,1 @@
-${CARGO_TARGET_DIR:-../target}/debug/quill --pem-file - neuron-manage 2313380519530470538 --disburse --stop-dissolving | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -
+${CARGO_TARGET_DIR:-../target}/debug/quill neuron-manage 2313380519530470538 --disburse --stop-dissolving --pem-file - | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -