+++ tests/commands/neuron-manage-spawn.sh
@@ -1,1 +1,1 @@
-${CARGO_TARGET_DIR:-../target}/debug/quill --pem-file - neuron-manage 2_313_380_519_530_470_538 --spawn | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -
+${CARGO_TARGET_DIR:-../target}/debug/quill neuron-manage 2_313_380_519_530_470_538 --spawn --pem-file - | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -