+++ tests/commands/neuron-manage-split.sh
@@ -1,1 +1,1 @@
-${CARGO_TARGET_DIR:-../target}/debug/quill --pem-file - neuron-manage 2313380519530470538 --split 100 | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -
+${CARGO_TARGET_DIR:-../target}/debug/quill neuron-manage 2313380519530470538 --split 100 --pem-file - | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -