+++ tests/commands/neuron-stake-nonce.sh
@@ -1,1 +1,1 @@
-cargo run -- --pem-file - neuron-stake --amount 12 --nonce 777 | cargo run -- send --dry-run -
+${CARGO_TARGET_DIR:-../target}/debug/quill neuron-stake --amount 12 --nonce 777 --pem-file - | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -