+++ tests/commands/neuron-stake-with-name.sh
@@ -1,1 +1,1 @@
-${CARGO_TARGET_DIR:-../target}/debug/quill --pem-file - neuron-stake --amount 12 --name myNeuron | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -
+${CARGO_TARGET_DIR:-../target}/debug/quill neuron-stake --amount 12 --name myNeuron --pem-file - | ${CARGO_TARGET_DIR:-../target}/debug/quill send --dry-run -